Remove the phone code, as agreed with Mike and Shaz.
The immediate motivation is that I want to move to the new version of CCI
(https://github.com/Microsoft/cci), which does not include the
ILGarbageCollect project, which is used by the phone plugin.

I removed everything phone-related (there were no unclear cases) and all
local variables left unused, but not any of the larger-scale extensibility
points that are no longer used.